Move maxPeriodText and minPeriodText to new location
