+++ app/pages/resource/ResourcePage.js
@@ -21,7 +21,8 @@ import PageWrapper from '../PageWrapper';
 import NotFoundPage from '../not-found/NotFoundPage';
 import ResourceCalendar from '../../shared/resource-calendar/ResourceCalendar';
 import injectT from '../../i18n/injectT';
-import { getMaxPeriodText, getResourcePageUrl, getMinPeriodText } from '../../utils/resourceUtils';
+import { getResourcePageUrl } from '../../utils/resourceUtils';
+import { getMaxPeriodText, getMinPeriodText } from '../../../src/domain/resource/utils';
 import { getEditReservationUrl } from '../../utils/reservationUtils';
 import ResourceHeader from './resource-header/ResourceHeader';
 import ResourceInfo from './resource-info/ResourceInfo';
