+++ app/pages/resource/reservation-info/ReservationInfo.js
@@ -7,7 +7,7 @@ import iconUser from 'hel-icons/dist/shapes/user-o.svg';
 import iconCalendar from '../../../assets/icons/calendar.svg';
 import iconClock from '../../../assets/icons/clock-o.svg';
 import WrappedText from '../../../shared/wrapped-text/WrappedText';
-import { getMaxPeriodText } from '../../../utils/resourceUtils';
+import { getMaxPeriodText } from '../../../../src/domain/resource/utils';
 import injectT from '../../../i18n/injectT';
 
 function renderLoginText(isLoggedIn, resource) {
