+++ app/pages/resource/resource-header/ResourceHeader.js
@@ -13,7 +13,8 @@ import injectT from '../../../i18n/injectT';
 import iconClock from '../../../assets/icons/clock-o.svg';
 import iconMap from '../../../assets/icons/map.svg';
 import FavoriteButton from '../../../shared/favorite-button/FavoriteButtonContainer';
-import { getHourlyPrice, getMaxPeriodText } from '../../../utils/resourceUtils';
+import { getHourlyPrice } from '../../../utils/resourceUtils';
+import { getMaxPeriodText } from '../../../../src/domain/resource/utils';
 
 function ResourceHeader({
   onBackClick,
