+++ app/utils/resourceUtils.js
@@ -103,24 +103,6 @@ function getHumanizedPeriod(period) {
   return `${moment.duration(period).hours()} h`;
 }
 
-function getMaxPeriodText(t, { maxPeriod }) {
-  const hours = moment.duration(maxPeriod).asHours();
-  const days = parseInt(moment.duration(maxPeriod).asDays(), 10);
-  if (days > 0) {
-    return t('ResourceHeader.maxPeriodDays', { days });
-  }
-  return t('ResourceHeader.maxPeriodHours', { hours });
-}
-
-function getMinPeriodText(t, { minPeriod }) {
-  const hours = moment.duration(minPeriod).asHours();
-  const days = parseInt(moment.duration(minPeriod).asDays(), 10);
-  if (days > 0) {
-    return t('ResourceHeader.minPeriodDays', { days });
-  }
-  return t('ResourceHeader.minPeriodHours', { hours });
-}
-
 function getOpeningHours(resource, selectedDate) {
   if (resource && resource.openingHours && resource.openingHours.length) {
     if (selectedDate) {
@@ -200,13 +182,11 @@ export {
   getAvailabilityDataForWholeDay,
   getHourlyPrice,
   getHumanizedPeriod,
-  getMaxPeriodText,
   getOpeningHours,
   getOpenReservations,
   getResourcePageUrl,
   getResourcePageUrlComponents,
   getTermsAndConditions,
   reservingIsRestricted,
-  getMinPeriodText,
   hasProducts,
 };
